requirements.txt
faker==22.6.0
pandas==2.1.4
databricks-sql-connector==2.9.3
python-dotenv==1.0.0
databricks-sdk==0.12.0